Create HiveMultiPool.py
This is the latest evolution of the script. It gives you a choice of No-Multiprocessing, Multiprocessing, and Multiprocessing with pool. I have also changed from using an Excell file to read and write the last_id to using a csv, which works much better. The columns for the CSV are as follows: Client_Name	Link_for_Console	Token	Status	Last_offense	NewOpen	Process_time	error_msg	Last_Process_Date